tests: pf: Fixed chflags problem in make_baseimg.sh.
make_baseimg.sh runs chflags -R noschg on the release directory in the
object tree.  That is done in order for rm -fr to work properly
afterwards.

chflags failing was misinterpreted as it failing to remove the flag.
But failure can also happen if the release directory does not exist.
In that case, running chflags should be skipped.

chflags is now only run if the release directory exists.  Failure of
the command now can clearly be interpreted as failure to remove the
flag.