tests: pf: Fail instead of skipping if VM not start.
vm_create() in pf_test_util.sh checks if vmctl.sh exits with status 2,
which means that the VM did not start or shut down after a short
while.  That would be interpreted as VM (bhyve) support not being
available on the system, and make the test skip.  However, that can
also happen if the image is unbootable or crashes.

The check has been changed to make the test fail instead of skipping,
since there can be multiple reasons for the failure.