tests: pf: Load nmdm when testing.
The tests start virtual machines when run, which need a console to
connect to.  The nmdm module is used for that.  Problems arise however
if it is not manually loaded before the tests start, such as virtual
machines failing to start without a good explanation.

The tests, that is pf_test_*.sh, have been fixed to load nmdm at the
start.