tests: pf: Renamed test scrub_forward6 to scrub6.
The implementation of the test scrub_forward6 was changed and does not
use forwarding anymore.  The name of the test however still contained
the word "forward", which can be misleading.

The name of the test has been changed to scrub6.  Related code has
also been changed to the new name.