tests: pf: Added comment in scrub6.py.
The file scrub6.py includes some code which is not currently activated
by the test, but can be used to generated test traffic.  This code
loads and stores test data from a file.  The purpose of the code could
be unclear to someone reading it, as the data is not used when this
code is not activated.

A comment has been added that helps clarify the purpose of this code.