tests: pf: Added new idea in "Future work" section of README.
Since the test scrub6 has been changed to generate traffic using
ping6, it can not check for payload integrity after reassembly in the
current implementation.  This could still be implemented in the
future, using inactive code that still exists in scrub6.py.

This possibility has been added as a new idea for future work in the
README file.