tests: pf: Added address prefix variable in test scrub_forward.
The test scrub_forward uses a common prefix for all IPv4 addresses
used.  Repeatedly writing the same address prefix throughout the test
can cause hard-to-find bugs if a typing error occurs.

The common prefix has now been assigned to a newly created variable,
and addresses are written as that variable plus a suffix.

For now, only scrub_forward has this change, but it can be applied to
the other tests as well.