tests: pf: No chroot in make_baseimg.sh.
As part of building the image file, make_baseimg.sh used to chroot
into it in order to install packages with pkg.  This showed however to
create varied problems with network connectivity, such as name
resolution (resolv.conf) not working properly.

Instead of running pkg under chroot, pkg is now run with the -r
option, which instructs it to install packages under the specified
root directory.