tests: pf: Added address prefix variables in all tests.
The tests use common prefixes for all IPv4 and IPv6 addresses used.
Repeatedly writing the same address prefixes throughout the tests can
cause hard-to-find bugs if a typing error occurs.

The common prefixes have now been assigned to two newly created
variables, and addresses are written as a variable plus a suffix.

This change was first implemented in the test scrub_forward, and it
has now been extended to the rest of the tests.