tests: pf: Changed idea in "Future work" section of README.
Since the test scrub has been changed to generate traffic using ping,
similarly to the test scrub6, it can not check for payload integrity
after reassembly in the current implementation.  This could still be
implemented in the future, using inactive code that still exists in
scrub.py, similarly to scrub6.py.

This possibility has been included in an idea for future work in the
README file.